Fix the spec on the TableHandler requirements
To be specific, it is clearly specified that:
 - no further calls shall not be made on handlers when they throw
   exceptions, and
 - calls of empty_physical_line shall occur after start_record and
   before end_record.

In addition, redundant descriptions are omitted.

The spec does not tell that start_record-end_record pairs are not
nested and neither are other pairs, but it seems OK because it is
already clear that they cannot be nested from their semantics;
that is, for example, records cannot resides in other records.